requirements.txt
llmauditor>=1.1.3
openai>=1.0.0  
langchain>=0.1.0
langchain-openai>=0.2.0
langchain-community>=0.0.10
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
rich>=13.0.0